requirements.txt
faker==37.1.0
langgraph==0.2.40
pandas==2.3.2
plotly==6.0.1
pyarrow==21.0.0
pymongo==4.11.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
scikit-learn==1.6.1
SQLAlchemy==2.0.39
streamlit==1.42.2
langchain==0.3.19
langchain-core==0.3.40
langchain-community==0.3.18
langchain-ollama==0.2.3
